organiza logica do componente crop
