+++ src/Components/Crop/crop.tsx
@@ -1,44 +1,27 @@
 import { X as CloseBtn } from "react-feather";
 
-import { ChangeEvent, ChangeEventHandler, useContext, useState } from "react";
+import { useContext } from "react";
 import FileContext from "../../Context/FileContext";
 
 import AvatarEditor from "react-avatar-editor";
+import useCrop from "./useCrop";
 
 function Crop() {
-  const [rotation, setRotation] = useState(0);
+  const { file, setFile, editor, setEditorRef, setUserProfilePic } =
+    useContext(FileContext);
 
   const {
-    file,
+    scaleValue,
+    rotation,
+    onScaleChange,
+    onRotationChange,
+    onCrop,
+    onClose,
+  } = useCrop({
     setFile,
     editor,
-    scaleValue,
-    setScaleValue,
-    setEditorRef,
     setUserProfilePic,
-  } = useContext(FileContext);
-
-  const onRotationChange = ({ target }: ChangeEvent<HTMLInputElement>) => {
-    setRotation(Number(target.value));
-  };
-
-  const onClose = () => {
-    setFile(null);
-  };
-
-  const onScaleChange = (scaleChangeEvent: ChangeEvent<HTMLInputElement>) => {
-    const scaleValue = parseFloat(scaleChangeEvent.target.value);
-    setScaleValue(scaleValue);
-  };
-
-  const onCrop = () => {
-    if (editor !== null) {
-      const url = editor.getImageScaledToCanvas().toDataURL();
-      setUserProfilePic(url);
-    }
-
-    onClose();
-  };
+  });
 
   if (!file) {
     return <></>;
@@ -61,14 +44,15 @@ function Crop() {
       </div>
       <div className="cropSettings">
         <div className="inputs">
-          <label>Zoom</label>
+          <label>Zoom {scaleValue}</label>
           <input
             style={{ width: "100%" }}
             type="range"
             value={scaleValue}
             name="points"
             min="1"
-            max="10"
+            max="3"
+            step="0.1"
             onChange={onScaleChange}
           />
         </div>
