+++ src/Components/Crop/useCrop.tsx
@@ -1,0 +1,47 @@
+import { ChangeEvent, useState } from "react";
+
+interface IUseCrop {
+  setFile: Function;
+  editor: any;
+  setUserProfilePic: Function;
+}
+
+function useCrop({ setFile, editor, setUserProfilePic }: IUseCrop) {
+  const [rotation, setRotation] = useState(0);
+  const [scaleValue, setScaleValue] = useState(1);
+
+  const onRotationChange = ({ target }: ChangeEvent<HTMLInputElement>) => {
+    setRotation(Number(target.value));
+  };
+
+  const onClose = () => {
+    setFile(null);
+  };
+
+  const onScaleChange = (scaleChangeEvent: ChangeEvent<HTMLInputElement>) => {
+    const scaleValue = parseFloat(scaleChangeEvent.target.value);
+    setScaleValue(scaleValue);
+  };
+
+  const onCrop = () => {
+    if (editor !== null) {
+      const url = editor.getImageScaledToCanvas().toDataURL();
+      setUserProfilePic(url);
+    }
+
+    onClose();
+  };
+
+  return {
+    scaleValue,
+    setScaleValue,
+    rotation,
+    setRotation,
+    onRotationChange,
+    onScaleChange,
+    onCrop,
+    onClose,
+  };
+}
+
+export default useCrop;
